FOFBApp: allow UInt32Chan to be used with Int32.
Assigning to a union member via pointers is potentially UB, even with
GCC extensions that allow type-punning via unions [1]. Avoid that by
using a temporary value. Removing the union entirely can also be
discussed.

This allows us to use the same set of functions to fill in asynInt32
values, which will be used for their sign-extensions functionality.

[1] https://gcc.gnu.org/onlinedocs/gcc/Optimize-Options.html#Type%2Dpunning