FOFBApp: add AlarmsAmp-Mon and -Cte records.
AlarmsAmp-Mon is needed as a single value to be checked for alarms,
while AlarmsAmp-Cte is used to label each bit that can be set in -Mon.

Since there isn't a straight-forward way to set the strings in the
waveform to their default values, we take advantage of the fact that
set_initial_parameters.py always has to be run for IOC bringup, allowing
us to set the values then and save then for subsequent runs.
Alternatives: INP field (available with Epics 7) [1] or the Sequencer
module (would add another module to our build just for this) [2].

[1] https://epics.anl.gov/tech-talk/2019/msg01696.php
[2] https://github.com/lnls-dig/dmm7510-epics-ioc/blob/4a3b2b211ff9baa73f904b5d35fdadda4ea10daa/dmm7510App/src/sncDCCT.stt#L17