FOFBApp: prepare driver for two FOFB_CC cores.
We launch only one IOC, but it's responsible for two sets of FOFB_CC
registers. In order to deal with this, we need to be ready for addr!=0
from these values and translate them to the appropriate places; in this
case, we also need to affect the coreID used in HALCS service name.