FOFBApp: add PINI to new -SP records.
From what can be understood, this makes it so the record wins a race
condition with the save_and_restore module and processes itself with the
saved value (therefore, after it's been written to the PV during
startup), wriiting them to hardware. Without this field, the record
wouldn't be processed on IOC startup, despite the saved value having
been written into it - this was detected by the fact that the -RB PV
didn't have the saved values.