FOFBApp: add new FOFB processing records.
The main addition from this commit is the high level OpMode
implementation, which allows us to safely switch between "fofb" and
"manual" modes for each power supply channel. It has two modes of
operation, one for each transition.

From "fofb" to "manual":
- pause accumulator
- copy the active effective current setpoint into the controller's
  current setpoint
- move to "closed_loop_manual" mode for the power supply

From "manual" to "fofb" (there's less that can be done in this
direction):
- move to "closed_loop_external" mode for the power supply
- resume accumulator

For implementation reasons, "pause accumulator" and "copy current
setpoint" always happen (it's easier than making the forward linking
conditional); "resume accumulator" needed a conditonal to make it a
no-op in the transition to "manual".

Finally. the power supply's mode is reported back to OpMode-Sts.

To enable this, we added the accumulator control records.

All records use FLNK and OUT parameters with CA (Channel Access) mode,
which seems to be necessary for at least some of them. This might be
related to records using others that are defined after them. We also use
CP for an INP field, which avoids having to refer to FOFB processing
related records in the RTM LAMP records.

And since we are here, move the inverse response matrix to this new
template file.

test-cp