FOFBApp: remove race condition from CopyCurrent.
It is possible that the value of CurrentRawRef-Mon hasn't been updated
after the accumulator is frozen, leading to the wrong value being
written to CurrentRaw-SP. We can work around this by adding a new
CurrentRawRefPassive record that will be triggered by the
PauseAccumulator helper and which will trigger CopyCurrent helper via
the CP input link. However, since CurrentRawRefPassive won't trigger an
input link when its value changes, it has to be skipped when going from
manual to fofb mode.

Since we are here, also change conditionals to compare OpMode to FOFB,
as that's easier to reason about.