FOFBApp: fix CopyCurrent.
The implementation for this record was relying on a current setpoint
value change triggering the input link using CP (it would be nice to
find some way to trigger links regardless of changes in value), in order
to avoid coupling the FOFBRtm.template records to the processing ones.
This leads to issues when the current isn't changing (communication isn't
synced up, gain is 0, among others), and even a risk of issues when the
current *is* changing, if it happens to land on the same value. Avoid
this entirely by using a FLNK in CurrentRawRefPassive.