FOFBApp: implement safe startup mode.
Restarting the IOC will zero out the power state and controller gain, in
order to avoid the possibility of big orbit kicks during testing. In the
future, for operation, we might disable PINI on even more output records
and instead add PINI to the input ones so we only read the state from
hardware and don't act on it.

These changes can be explained as:

- removing PwrState-Sel from .req so we always write 0
- removing CurrLoopMode-Sel from .req so we always write 0 as well (this
  has issues in the current implementation where something is causing
  the OpModeHelperSetPSMode record is getting triggered and setting
  CurrLoopMode-Sel anyway
- remove erroneous PINI from CurrentRawRefPassive and OpMode-Sts
- remove PINI from OpMode-Sel (though it's currently getting applied
  anyway)
- set FOFBAccGain-SP to 0 by default